requirementsdde.txt
asgiref==3.7.2
crispy-bootstrap5==2024.2
Django==5.0.2
django-ckeditor-5==0.2.12
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-filter==24.2
django-js-asset==2.2.0
djangorestframework==3.15.1
environs==10.3.0
Markdown==3.6
marshmallow==3.21.0
packaging==23.2
pillow==10.2.0
python-dotenv==1.0.1
slugify==0.0.1
sqlparse==0.4.4
tzdata==2024.1
Unidecode==1.3.8
